requirements.txt
Django==1.11
django-taggit==1.8
